Fix: PayFirma gateway validates Credit Card even if not selected or active
As written, PayFirma will block the checkout process and ask the user to enter their credit card information even if the user has selected a different payment method (Cheque, COD, PayPal etc.)

Further, this validation runs even if the PayFirma plugin is deactivated in WooCommerce (using the active switch). This means it will block ALL payments on a website if it's active EXCEPT PayFirma payments. The only way to stop this is to deactivate the plugin as a whole (e.g. remove it from the site entirely.)

This is because the woocommerce_checkout_process hook runs every time a user submits an order, and this code placed where it is means it will run always (when it should not be.)

The proposed fix checks that the user actually selected PayFirma as the payment method before validating the credit card information. This also takes care of ensuring PayFirma is active in WooCommerce as well, because if it wasn't active then $_POST['payment_method'] will never be 'payfirma_gateway'.